pyproject.toml
[project]
name = "billing"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "sqlmodel>=0.0.24",
    "asyncpg>=0.30.0",
    "pyyaml>=6.0.2",
    "redis>=5.2.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    "black>=24.10.0",
    "isort>=5.13.2",
    "aiosqlite>=0.20.0",
    "greenlet>=3.3.0",
    "reportlab>=4.0.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
